+++ src/plugins/data_source_management/public/components/data_source_table/data_source_table.tsx
@@ -9,18 +9,17 @@ import {
   EuiConfirmModal,
   EuiFlexGroup,
   EuiFlexItem,
-  EuiGlobalToastList,
-  EuiGlobalToastListToast,
   EuiInMemoryTable,
   EuiPageContent,
+  EuiPanel,
   EuiSpacer,
   EuiText,
   EuiTitle,
 } from '@elastic/eui';
 import React, { useState } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import { FormattedMessage } from '@osd/i18n/react';
 import { useEffectOnce } from 'react-use';
+import { i18n } from '@osd/i18n';
 import { getListBreadcrumbs } from '../breadcrumbs';
 import {
   reactRouterNavigate,
@@ -31,16 +30,17 @@ import { CreateButton } from '../create_button';
 import { deleteMultipleDataSources, getDataSources } from '../utils';
 import { LoadingMask } from '../loading_mask';
 import {
-  cancelText,
-  deleteText,
-  dsListingAriaRegion,
-  dsListingDeleteDataSourceConfirmation,
-  dsListingDeleteDataSourceDescription,
-  dsListingDeleteDataSourceTitle,
-  dsListingDeleteDataSourceWarning,
-  dsListingDescription,
-  dsListingPageTitle,
-  dsListingTitle,
+  CANCEL_TEXT,
+  DELETE_TEXT,
+  DS_LISTING_ARIA_REGION,
+  DS_LISTING_DATA_SOURCE_DELETE_ACTION,
+  DS_LISTING_DATA_SOURCE_DELETE_IMPACT,
+  DS_LISTING_DATA_SOURCE_DELETE_WARNING,
+  DS_LISTING_DATA_SOURCE_MULTI_DELETE_TITLE,
+  DS_LISTING_DESCRIPTION,
+  DS_LISTING_NO_DATA,
+  DS_LISTING_PAGE_TITLE,
+  DS_LISTING_TITLE,
 } from '../text_content/text_content';
 
 /* Table config */
@@ -56,17 +56,17 @@ const sorting = {
   },
 };
 
-const toastLifeTimeMs = 6000;
-
 export const DataSourceTable = ({ history }: RouteComponentProps) => {
-  const { chrome, setBreadcrumbs, savedObjects } = useOpenSearchDashboards<
-    DataSourceManagementContext
-  >().services;
+  const {
+    chrome,
+    setBreadcrumbs,
+    savedObjects,
+    notifications: { toasts },
+  } = useOpenSearchDashboards<DataSourceManagementContext>().services;
 
   /* Component state variables */
   const [dataSources, setDataSources] = useState<DataSourceTableItem[]>([]);
   const [selectedDataSources, setSelectedDataSources] = useState<DataSourceTableItem[]>([]);
-  const [toasts, setToasts] = React.useState<EuiGlobalToastListToast[]>([]);
   const [isLoading, setIsLoading] = React.useState<boolean>(false);
   const [isDeleting, setIsDeleting] = React.useState<boolean>(false);
   const [confirmDeleteVisible, setConfirmDeleteVisible] = React.useState(false);
@@ -77,7 +77,7 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
     setBreadcrumbs(getListBreadcrumbs());
 
     /* Browser - Page Title */
-    chrome.docTitle.change(dsListingPageTitle);
+    chrome.docTitle.change(DS_LISTING_PAGE_TITLE);
 
     /* fetch data sources*/
     fetchDataSources();
@@ -93,10 +93,7 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
         setDataSources([]);
         handleDisplayToastMessage({
           id: 'dataSourcesManagement.dataSourceListing.fetchDataSourceFailMsg',
-          defaultMessage:
-            'Error occurred while fetching the records for Data sources. Please try it again',
-          color: 'warning',
-          iconType: 'alert',
+          defaultMessage: 'Error occurred while fetching the records for Data sources.',
         });
       })
       .finally(() => {
@@ -109,14 +106,13 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
     return (
       <EuiButton
         color="danger"
-        iconType="trash"
         onClick={() => {
           setConfirmDeleteVisible(true);
         }}
         data-test-subj="deleteDataSourceConnections"
         disabled={selectedDataSources.length === 0}
       >
-        Delete {selectedDataSources.length || ''} connection
+        Delete {selectedDataSources.length || ''} {selectedDataSources.length ? 'connection' : ''}
         {selectedDataSources.length >= 2 ? 's' : ''}
       </EuiButton>
     );
@@ -180,7 +176,7 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
   const tableRenderDeleteModal = () => {
     return confirmDeleteVisible ? (
       <EuiConfirmModal
-        title={dsListingDeleteDataSourceTitle}
+        title={DS_LISTING_DATA_SOURCE_MULTI_DELETE_TITLE}
         onCancel={() => {
           setConfirmDeleteVisible(false);
         }}
@@ -188,13 +184,13 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
           setConfirmDeleteVisible(false);
           onClickDelete();
         }}
-        cancelButtonText={cancelText}
-        confirmButtonText={deleteText}
+        cancelButtonText={CANCEL_TEXT}
+        confirmButtonText={DELETE_TEXT}
         defaultFocusedButton="confirm"
       >
-        <p>{dsListingDeleteDataSourceDescription}</p>
-        <p>{dsListingDeleteDataSourceConfirmation}</p>
-        <p>{dsListingDeleteDataSourceWarning}</p>
+        <p>{DS_LISTING_DATA_SOURCE_DELETE_ACTION}</p>
+        <p>{DS_LISTING_DATA_SOURCE_DELETE_IMPACT}</p>
+        <p>{DS_LISTING_DATA_SOURCE_DELETE_WARNING}</p>
       </EuiConfirmModal>
     ) : null;
   };
@@ -214,9 +210,7 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
         handleDisplayToastMessage({
           id: 'dataSourcesManagement.dataSourceListing.deleteDataSourceFailMsg',
           defaultMessage:
-            'Error occurred while deleting few/all selected records for Data sources. Please try it again',
-          color: 'warning',
-          iconType: 'alert',
+            'Error occurred while deleting selected records for Data sources. Please try it again',
         });
       })
       .finally(() => {
@@ -234,21 +228,13 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
   };
 
   /* Toast Handlers */
-  const removeToast = (id: string) => {
-    setToasts(toasts.filter((toast) => toast.id !== id));
-  };
 
-  const handleDisplayToastMessage = ({ id, defaultMessage, color, iconType }: ToastMessageItem) => {
-    const failureMsg = <FormattedMessage id={id} defaultMessage={defaultMessage} />;
-    setToasts([
-      ...toasts,
-      {
-        title: failureMsg,
-        id: failureMsg.props.id,
-        color,
-        iconType,
-      },
-    ]);
+  const handleDisplayToastMessage = ({ id, defaultMessage }: ToastMessageItem) => {
+    toasts.addWarning(
+      i18n.translate(id, {
+        defaultMessage,
+      })
+    );
   };
 
   /* Render Ui elements*/
@@ -261,11 +247,11 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
       <EuiFlexGroup justifyContent="spaceBetween">
         <EuiFlexItem grow={false}>
           <EuiTitle>
-            <h2>{dsListingTitle}</h2>
+            <h2>{DS_LISTING_TITLE}</h2>
           </EuiTitle>
           <EuiSpacer size="s" />
           <EuiText>
-            <p>{dsListingDescription}</p>
+            <p>{DS_LISTING_DESCRIPTION}</p>
           </EuiText>
         </EuiFlexItem>
         <EuiFlexItem grow={false}>{createButton}</EuiFlexItem>
@@ -277,10 +263,44 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
   const renderTableContent = () => {
     return (
       <>
+        {/* Data sources table*/}
+        <EuiInMemoryTable
+          allowNeutralSort={false}
+          itemId="id"
+          isSelectable={true}
+          selection={selection}
+          items={dataSources}
+          columns={columns}
+          pagination={pagination}
+          sorting={sorting}
+          search={search}
+          loading={isLoading}
+        />
+      </>
+    );
+  };
+
+  const renderEmptyState = () => {
+    return (
+      <>
+        <EuiSpacer size="l" />
+        <EuiPanel hasBorder={false} hasShadow={false} style={{ textAlign: 'center' }}>
+          <EuiText>{DS_LISTING_NO_DATA}</EuiText>
+          <EuiSpacer />
+          {createButton}
+        </EuiPanel>
+        <EuiSpacer size="l" />
+      </>
+    );
+  };
+
+  const renderContent = () => {
+    return (
+      <>
         <EuiPageContent
           data-test-subj="dataSourceTable"
           role="region"
-          aria-label={dsListingAriaRegion}
+          aria-label={DS_LISTING_ARIA_REGION}
         >
           {/* Header */}
           {renderHeader()}
@@ -290,46 +310,16 @@ export const DataSourceTable = ({ history }: RouteComponentProps) => {
           {/* Delete confirmation modal*/}
           {tableRenderDeleteModal()}
 
-          {/* Data sources table*/}
-          <EuiInMemoryTable
-            allowNeutralSort={false}
-            itemId="id"
-            isSelectable={true}
-            selection={selection}
-            items={dataSources}
-            columns={columns}
-            pagination={pagination}
-            sorting={sorting}
-            search={search}
-            loading={isLoading}
-          />
+          {!isLoading && (!dataSources || !dataSources.length)
+            ? renderEmptyState()
+            : renderTableContent()}
         </EuiPageContent>
         {isDeleting ? <LoadingMask /> : null}
       </>
     );
   };
 
-  const renderContent = () => {
-    return (
-      <>
-        {renderTableContent()}
-        {}
-      </>
-    );
-  };
-
-  return (
-    <>
-      {renderContent()}
-      <EuiGlobalToastList
-        toasts={toasts}
-        dismissToast={({ id }) => {
-          removeToast(id);
-        }}
-        toastLifeTimeMs={toastLifeTimeMs}
-      />
-    </>
-  );
+  return renderContent();
 };
 
 export const DataSourceTableWithRouter = withRouter(DataSourceTable);