+++ src/plugins/data_source_management/public/components/validation/datasource_form_validation.ts
@@ -4,19 +4,10 @@
  */
 
 import { isValidUrl } from '../utils';
+import { DATA_SOURCE_VALIDATION_TITLE_EXISTS } from '../text_content';
 import { CreateDataSourceState } from '../create_data_source_wizard/components/create_form/create_data_source_form';
-import { AuthType } from '../../types';
 import { EditDataSourceState } from '../edit_data_source/components/edit_form/edit_data_source_form';
-import { UpdatePasswordFormType } from '../../types';
-import {
-  dataSourceValidationEndpointNotValid,
-  dataSourceValidationNewPasswordEmpty,
-  dataSourceValidationNoPasswordMatch,
-  dataSourceValidationOldPasswordEmpty,
-  dataSourceValidationPasswordEmpty,
-  dataSourceValidationTitleEmpty,
-  dataSourceValidationUsernameEmpty,
-} from '../text_content';
+import { AuthType } from '../../types';
 
 export interface CreateEditDataSourceValidation {
   title: string[];
@@ -27,12 +18,6 @@ export interface CreateEditDataSourceValidation {
   };
 }
 
-export interface UpdatePasswordValidation {
-  oldPassword: string[];
-  newPassword: string[];
-  confirmNewPassword: string[];
-}
-
 export const defaultValidation: CreateEditDataSourceValidation = {
   title: [],
   endpoint: [],
@@ -41,34 +26,46 @@ export const defaultValidation: CreateEditDataSourceValidation = {
     password: [],
   },
 };
-export const defaultPasswordValidationByField: UpdatePasswordValidation = {
-  oldPassword: [],
-  newPassword: [],
-  confirmNewPassword: [],
+
+export const isTitleValid = (
+  title: string,
+  existingDatasourceNamesList: string[],
+  existingTitle: string
+) => {
+  const isValid = {
+    valid: true,
+    error: '',
+  };
+  /* Title validation */
+  if (!title?.trim?.().length) {
+    isValid.valid = false;
+  } else if (
+    title.toLowerCase() !== existingTitle.toLowerCase() &&
+    Array.isArray(existingDatasourceNamesList) &&
+    existingDatasourceNamesList.includes(title.toLowerCase())
+  ) {
+    /* title already exists */
+    isValid.valid = false;
+    isValid.error = DATA_SOURCE_VALIDATION_TITLE_EXISTS;
+  }
+  return isValid;
 };
 
 export const performDataSourceFormValidation = (
-  formValues: CreateDataSourceState | EditDataSourceState
+  formValues: CreateDataSourceState | EditDataSourceState,
+  existingDatasourceNamesList: string[],
+  existingTitle: string
 ) => {
-  const validationByField: CreateEditDataSourceValidation = {
-    title: [],
-    endpoint: [],
-    createCredential: {
-      username: [],
-      password: [],
-    },
-  };
-  const formErrorMessages: string[] = [];
   /* Title validation */
-  if (!formValues?.title?.trim?.().length) {
-    validationByField.title.push(dataSourceValidationTitleEmpty);
-    formErrorMessages.push(dataSourceValidationTitleEmpty);
+  const titleValid = isTitleValid(formValues?.title, existingDatasourceNamesList, existingTitle);
+
+  if (!titleValid.valid) {
+    return false;
   }
 
   /* Endpoint Validation */
   if (!isValidUrl(formValues?.endpoint)) {
-    validationByField.endpoint.push(dataSourceValidationEndpointNotValid);
-    formErrorMessages.push(dataSourceValidationEndpointNotValid);
+    return false;
   }
 
   /* Credential Validation */
@@ -77,46 +74,14 @@ export const performDataSourceFormValidation = (
   if (formValues?.auth?.type === AuthType.UsernamePasswordType) {
     /* Username */
     if (!formValues.auth.credentials?.username) {
-      validationByField.createCredential.username.push(dataSourceValidationUsernameEmpty);
-      formErrorMessages.push(dataSourceValidationUsernameEmpty);
+      return false;
     }
 
     /* password */
     if (!formValues.auth.credentials?.password) {
-      validationByField.createCredential.password.push(dataSourceValidationPasswordEmpty);
-      formErrorMessages.push(dataSourceValidationPasswordEmpty);
+      return false;
     }
   }
 
-  return {
-    formErrors: formErrorMessages,
-    formErrorsByField: { ...validationByField },
-  };
-};
-
-export const validateUpdatePassword = (passwords: UpdatePasswordFormType) => {
-  const validationByField: UpdatePasswordValidation = {
-    oldPassword: [],
-    newPassword: [],
-    confirmNewPassword: [],
-  };
-
-  const formErrorMessages: string[] = [];
-
-  if (!passwords.oldPassword) {
-    validationByField.oldPassword.push(dataSourceValidationOldPasswordEmpty);
-    formErrorMessages.push(dataSourceValidationOldPasswordEmpty);
-  }
-  if (!passwords.newPassword) {
-    validationByField.newPassword.push(dataSourceValidationNewPasswordEmpty);
-    formErrorMessages.push(dataSourceValidationNewPasswordEmpty);
-  } else if (passwords.confirmNewPassword !== passwords.newPassword) {
-    validationByField.confirmNewPassword.push(dataSourceValidationNoPasswordMatch);
-    formErrorMessages.push(dataSourceValidationNoPasswordMatch);
-  }
-
-  return {
-    formValidationErrors: formErrorMessages,
-    formValidationErrorsByField: { ...validationByField },
-  };
+  return true;
 };